- added tests-diag patch (try to debug test failures on builder)

+++ m4.spec
@@ -16,6 +16,7 @@ Group:		Applications/Text
 Source0:	https://ftp.gnu.org/gnu/m4/%{name}-%{version}.tar.xz
 # Source0-md5:	0d90823e1426f1da2fd872df0311298d
 Patch0:		%{name}-info.patch
+Patch1:		%{name}-tests-diag.patch
 URL:		http://www.gnu.org/software/m4/
 BuildRequires:	gettext-tools >= 0.19.2
 BuildRequires:	tar >= 1:1.22
@@ -75,6 +76,7 @@ Statyczna biblioteka m4.
 %prep
 %setup -q
 %patch0 -p1
+%patch1 -p1
 
 %{__sed} -i '1 i @documentencoding ISO-8859-1' doc/m4.texi
 
@@ -93,7 +95,14 @@ touch checks/stamp-checks
 
 %{__make}
 
-%{?with_tests:%{__make} check || cat tests/test-suite.log}
+%if %{with tests}
+st=0
+%{__make} check || st=1
+if [ "$st" -ne 0 ]; then
+	cat tests/test-suite.log
+	exit 1
+fi
+%endif
 
 %install
 rm -rf $RPM_BUILD_ROOT
